requirements.txt
torch
torchvision
timm
wandb
python-dotenv
PyYAML
opencv-python
scikit-learn
matplotlib
seaborn
pillow
onnx